Update category CRUD after merging main

+++ CFMS_Group6/src/main/java/controller/category/DeleteCategoryController.java
@@ -2,7 +2,6 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-
 package controller.category;
 
 import dal.CategoryDao;
@@ -13,25 +12,28 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.sql.SQLException;
 
 /**
  *
  * @author quang
  */
-@WebServlet(name="DeleteCategory", urlPatterns={"/category/DeleteCategory"})
-public class DeleteCategory extends HttpServlet {
-   
-    /** 
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code> methods.
+@WebServlet(name = "DeleteCategoryController", urlPatterns = {"/category/DeleteCategoryController"})
+public class DeleteCategoryController extends HttpServlet {
+
+    /**
+     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
+     * methods.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
-
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /** 
+    /**
      * Handles the HTTP <code>GET</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -39,16 +41,29 @@ public class DeleteCategory extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
+        String status;
         int id = Integer.parseInt(request.getParameter("id"));
         CategoryDao cDao = new CategoryDao();
-        cDao.deleteCategory(id);
-        response.sendRedirect("ViewCategory");
-       
-    } 
+        try {
+            if (cDao.isAssetInCategoryEmpty(id)) {
+                cDao.deleteCategory(id);
+            }
+            else{
+                status = "Lỗi: Danh mục đang chứa các tài sản khác.";
+                request.setAttribute("status", status);
+            }
+        } catch (ClassNotFoundException | SQLException e) {
+            
+        }
 
-    /** 
+        response.sendRedirect("ViewCategoryController");
+
+    }
+
+    /**
      * Handles the HTTP <code>POST</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -56,12 +71,13 @@ public class DeleteCategory extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
-        
+            throws ServletException, IOException {
+
     }
 
-    /** 
+    /**
      * Returns a short description of the servlet.
+     *
      * @return a String containing servlet description
      */
     @Override
